Share code between front/front-api
