+++ front-api/routes/w/[wId]/credits/metronome-balances.ts
@@ -1,18 +1,39 @@
-import { metronomeBalanceToDisplayData } from "@app/lib/api/credits/metronome_balances";
-import { listMetronomeBalances } from "@app/lib/metronome/client";
-import { getCreditTypeProgrammaticUsdId } from "@app/lib/metronome/constants";
-import { isMetronomeExcessCredit } from "@app/lib/metronome/types";
-import type {
-  CreditDisplayData,
-  GetCreditsResponseBody,
-} from "@app/types/credits";
+import type { MetronomeBalancesError } from "@app/lib/api/credits/metronome_balances";
+import { getMetronomeBalances } from "@app/lib/api/credits/metronome_balances";
+import type { GetCreditsResponseBody } from "@app/types/credits";
+import { assertNever } from "@app/types/shared/utils/assert_never";
 import { workspaceApp } from "@front-api/middlewares/ctx";
+import type { HandlerResult } from "@front-api/middlewares/utils";
 import { apiError } from "@front-api/middlewares/utils";
+import type { Context } from "hono";
+
+function balancesErrorToApi(ctx: Context, err: MetronomeBalancesError) {
+  switch (err.type) {
+    case "not_configured":
+      return apiError(ctx, {
+        status_code: 400,
+        api_error: {
+          type: "invalid_request_error",
+          message: "Workspace is not configured for Metronome billing.",
+        },
+      });
+    case "balances_fetch_failed":
+      return apiError(ctx, {
+        status_code: 500,
+        api_error: {
+          type: "internal_server_error",
+          message: `Failed to retrieve Metronome balances: ${err.cause?.message ?? ""}`,
+        },
+      });
+    default:
+      assertNever(err.type);
+  }
+}
 
 // Mounted at /api/w/:wId/credits/metronome-balances.
 const app = workspaceApp();
 
-app.get("/", async (ctx) => {
+app.get("/", async (ctx): HandlerResult<GetCreditsResponseBody> => {
   const auth = ctx.get("auth");
 
   if (!auth.isAdmin()) {
@@ -26,40 +47,11 @@ app.get("/", async (ctx) => {
     });
   }
 
-  const workspace = auth.getNonNullableWorkspace();
-  const { metronomeCustomerId } = workspace;
-  if (!metronomeCustomerId) {
-    return apiError(ctx, {
-      status_code: 400,
-      api_error: {
-        type: "invalid_request_error",
-        message: "Workspace is not configured for Metronome billing.",
-      },
-    });
-  }
-
-  const result = await listMetronomeBalances(metronomeCustomerId);
+  const result = await getMetronomeBalances(auth);
   if (result.isErr()) {
-    return apiError(ctx, {
-      status_code: 500,
-      api_error: {
-        type: "internal_server_error",
-        message: `Failed to retrieve Metronome balances: ${result.error.message}`,
-      },
-    });
+    return balancesErrorToApi(ctx, result.error);
   }
-
-  const programmaticUsdCreditTypeId = getCreditTypeProgrammaticUsdId();
-  const credits: CreditDisplayData[] = result.value
-    .filter(
-      (entry) =>
-        entry.access_schedule?.credit_type?.id ===
-          programmaticUsdCreditTypeId && !isMetronomeExcessCredit(entry)
-    )
-    .map(metronomeBalanceToDisplayData);
-
-  const body: GetCreditsResponseBody = { credits };
-  return ctx.json(body);
+  return ctx.json(result.value);
 });
 
 export default app;
