+++ front/lib/api/credits/awu_pool_summary.ts
@@ -10,9 +10,8 @@ import {
   getProductSeatTypes,
   getSeatTypesByProductIdFromContract,
 } from "@app/lib/metronome/seat_types";
-import { apiError } from "@app/logger/withlogging";
-import type { WithAPIErrorResponse } from "@app/types/error";
-import type { NextApiRequest, NextApiResponse } from "next";
+import type { Result } from "@app/types/shared/result";
+import { Err, Ok } from "@app/types/shared/result";
 
 export type AwuPoolSummaryResponseBody = {
   totalRemainingCredits: number;
@@ -20,142 +19,115 @@ export type AwuPoolSummaryResponseBody = {
   resetDate: string;
 };
 
-export async function handleAwuPoolSummaryRequest(
-  req: NextApiRequest,
-  res: NextApiResponse<WithAPIErrorResponse<AwuPoolSummaryResponseBody>>,
+export class AwuPoolSummaryError extends Error {
+  constructor(
+    readonly type:
+      | "not_configured"
+      | "balances_fetch_failed"
+      | "invoices_fetch_failed",
+    readonly cause?: Error
+  ) {
+    super(type);
+  }
+}
+
+export async function getAwuPoolSummary(
   auth: Authenticator
-): Promise<void> {
-  if (!auth.isAdmin()) {
-    return apiError(req, res, {
-      status_code: 403,
-      api_error: {
-        type: "workspace_auth_error",
-        message:
-          "Only users that are `admins` for the current workspace can view credits.",
-      },
+): Promise<Result<AwuPoolSummaryResponseBody, AwuPoolSummaryError>> {
+  const workspace = auth.getNonNullableWorkspace();
+  const subscription = auth.subscription();
+  const { metronomeCustomerId } = workspace;
+  if (!metronomeCustomerId || !subscription?.metronomeContractId) {
+    return new Err(new AwuPoolSummaryError("not_configured"));
+  }
+  const { metronomeContractId } = subscription;
+
+  const [balancesResult, invoicesResult] = await Promise.all([
+    listMetronomeBalances(metronomeCustomerId),
+    listMetronomeDraftInvoices(metronomeCustomerId),
+  ]);
+
+  if (balancesResult.isErr()) {
+    return new Err(
+      new AwuPoolSummaryError("balances_fetch_failed", balancesResult.error)
+    );
+  }
+  if (invoicesResult.isErr()) {
+    return new Err(
+      new AwuPoolSummaryError("invoices_fetch_failed", invoicesResult.error)
+    );
+  }
+
+  const now = Date.now();
+
+  // Find the canonical billing period end from the current draft invoice.
+  const currentInvoice = invoicesResult.value.find((inv) => {
+    if (inv.contract_id !== metronomeContractId) {
+      return false;
+    }
+    if (!inv.start_timestamp || !inv.end_timestamp) {
+      return false;
+    }
+    const startMs = new Date(inv.start_timestamp).getTime();
+    const endMs = new Date(inv.end_timestamp).getTime();
+    return startMs <= now && now < endMs;
+  });
+
+  if (!currentInvoice?.start_timestamp || !currentInvoice.end_timestamp) {
+    return new Ok({
+      totalRemainingCredits: 0,
+      totalActiveCredits: 0,
+      resetDate: "",
     });
   }
 
-  switch (req.method) {
-    case "GET": {
-      const workspace = auth.getNonNullableWorkspace();
-      const subscription = auth.subscription();
-      const { metronomeCustomerId } = workspace;
-      if (!metronomeCustomerId || !subscription?.metronomeContractId) {
-        return apiError(req, res, {
-          status_code: 400,
-          api_error: {
-            type: "invalid_request_error",
-            message: "Workspace is not configured for Metronome billing.",
-          },
-        });
+  const resetDate = ceilToMidnightUTC(
+    new Date(currentInvoice.end_timestamp)
+  ).toISOString();
+
+  // Filter to active, non-seat AWU pool credits and commits. The set of
+  // seat product IDs is derived from the contract's tagged subscriptions
+  // (via the `DUST_SEAT_TYPE` custom field) rather than a hardcoded list.
+  // The contract filter prevents sandbox prepaid commits on other contracts
+  // from inflating the balance.
+  const awuCreditTypeId = getCreditTypeAwuId();
+  const activeContract = await getActiveContract(workspace.sId);
+  const productSeatTypes = await getProductSeatTypes();
+  const seatProductIds = activeContract
+    ? new Set(
+        getSeatTypesByProductIdFromContract(
+          activeContract,
+          productSeatTypes
+        ).keys()
+      )
+    : new Set<string>();
+  const awuBalances = balancesResult.value.filter(
+    (entry) =>
+      entry.access_schedule?.credit_type?.id === awuCreditTypeId &&
+      !seatProductIds.has(entry.product.id) &&
+      (entry.contract?.id === metronomeContractId || !entry.contract)
+  );
+
+  let totalRemainingCredits = 0;
+  let totalActiveCredits = 0;
+  for (const entry of awuBalances) {
+    const scheduleItems = entry.access_schedule?.schedule_items ?? [];
+    const isActive = scheduleItems.some((item) => {
+      const itemStartMs = new Date(item.starting_at).getTime();
+      const itemEndMs = new Date(item.ending_before).getTime();
+      return itemStartMs <= now && now < itemEndMs;
+    });
+    if (isActive) {
+      totalRemainingCredits += entry.balance ?? 0;
+      for (const item of scheduleItems) {
+        totalActiveCredits += item.amount;
       }
-      const { metronomeContractId } = subscription;
-
-      const [balancesResult, invoicesResult] = await Promise.all([
-        listMetronomeBalances(metronomeCustomerId),
-        listMetronomeDraftInvoices(metronomeCustomerId),
-      ]);
-
-      if (balancesResult.isErr()) {
-        return apiError(req, res, {
-          status_code: 500,
-          api_error: {
-            type: "internal_server_error",
-            message: `Failed to retrieve Metronome balances: ${balancesResult.error.message}`,
-          },
-        });
-      }
-      if (invoicesResult.isErr()) {
-        return apiError(req, res, {
-          status_code: 500,
-          api_error: {
-            type: "internal_server_error",
-            message: `Failed to retrieve Metronome invoices: ${invoicesResult.error.message}`,
-          },
-        });
-      }
-
-      const now = Date.now();
-
-      // Find the canonical billing period end from the current draft invoice.
-      const currentInvoice = invoicesResult.value.find((inv) => {
-        if (inv.contract_id !== metronomeContractId) {
-          return false;
-        }
-        if (!inv.start_timestamp || !inv.end_timestamp) {
-          return false;
-        }
-        const startMs = new Date(inv.start_timestamp).getTime();
-        const endMs = new Date(inv.end_timestamp).getTime();
-        return startMs <= now && now < endMs;
-      });
-
-      if (!currentInvoice?.start_timestamp || !currentInvoice.end_timestamp) {
-        return res.status(200).json({
-          totalRemainingCredits: 0,
-          totalActiveCredits: 0,
-          resetDate: "",
-        });
-      }
-
-      const resetDate = ceilToMidnightUTC(
-        new Date(currentInvoice.end_timestamp)
-      ).toISOString();
-
-      // Filter to active, non-seat AWU pool credits and commits. The set of
-      // seat product IDs is derived from the contract's tagged subscriptions
-      // (via the `DUST_SEAT_TYPE` custom field) rather than a hardcoded list.
-      // The contract filter prevents sandbox prepaid commits on other contracts
-      // from inflating the balance.
-      const awuCreditTypeId = getCreditTypeAwuId();
-      const activeContract = await getActiveContract(workspace.sId);
-      const productSeatTypes = await getProductSeatTypes();
-      const seatProductIds = activeContract
-        ? new Set(
-            getSeatTypesByProductIdFromContract(
-              activeContract,
-              productSeatTypes
-            ).keys()
-          )
-        : new Set<string>();
-      const awuBalances = balancesResult.value.filter(
-        (entry) =>
-          entry.access_schedule?.credit_type?.id === awuCreditTypeId &&
-          !seatProductIds.has(entry.product.id) &&
-          (entry.contract?.id === metronomeContractId || !entry.contract)
-      );
-
-      let totalRemainingCredits = 0;
-      let totalActiveCredits = 0;
-      for (const entry of awuBalances) {
-        const scheduleItems = entry.access_schedule?.schedule_items ?? [];
-        const isActive = scheduleItems.some((item) => {
-          const itemStartMs = new Date(item.starting_at).getTime();
-          const itemEndMs = new Date(item.ending_before).getTime();
-          return itemStartMs <= now && now < itemEndMs;
-        });
-        if (isActive) {
-          totalRemainingCredits += entry.balance ?? 0;
-          for (const item of scheduleItems) {
-            totalActiveCredits += item.amount;
-          }
-        }
-      }
-
-      return res.status(200).json({
-        totalRemainingCredits,
-        totalActiveCredits,
-        resetDate,
-      });
     }
-    default:
-      return apiError(req, res, {
-        status_code: 405,
-        api_error: {
-          type: "method_not_supported_error",
-          message: "The method passed is not supported, GET is expected.",
-        },
-      });
   }
+
+  return new Ok({
+    totalRemainingCredits,
+    totalActiveCredits,
+    resetDate,
+  });
 }
